% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empiricalLikelihood.R
\name{empiricalLikelihood}
\alias{empiricalLikelihood}
\title{Compute the empirical likelihood for data: data
with lag: lag}
\usage{
empiricalLikelihood(data, lag = 1)
}
\arguments{
\item{data}{the data to construct from}

\item{lag}{the lag in data to use}
}
\description{
Compute the empirical likelihood for data: data
with lag: lag
}
