% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model0_R}
\alias{model0_R}
\title{Model0
The first model, a toy example.
y_{i+1} = y_{i} + N(theta, 1)}
\usage{
model0_R(N = 50, M = 1, theta)
}
\arguments{
\item{N}{the length of the output data}

\item{M}{the number of realisations}

\item{theta}{the model parameter}
}
\description{
Model0
The first model, a toy example.
y_{i+1} = y_{i} + N(theta, 1)
}
